Refactored application comments component
